Adds generic RegistrationForm component, adds 2023 hackathon mailchimp form based on 2021 implementation
